# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * GuardianBridge Lite telemetry and structural health monitoring API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
}

export type Scenario = typeof Scenario[keyof typeof Scenario];


export const Scenario = {
  Normal_Operation: 'Normal Operation',
  Increased_Vibration: 'Increased Vibration',
  Loose_Joint: 'Loose Joint',
  Excess_Load: 'Excess Load',
} as const;

export type TelemetryStatus = typeof TelemetryStatus[keyof typeof TelemetryStatus];


export const TelemetryStatus = {
  NORMAL: 'NORMAL',
  WARNING: 'WARNING',
  ANOMALY: 'ANOMALY',
  CRITICAL: 'CRITICAL',
} as const;

export type EventSeverity = typeof EventSeverity[keyof typeof EventSeverity];


export const EventSeverity = {
  NORMAL: 'NORMAL',
  WARNING: 'WARNING',
  CRITICAL: 'CRITICAL',
} as const;

export type LoraStatus = typeof LoraStatus[keyof typeof LoraStatus];


export const LoraStatus = {
  CONNECTED: 'CONNECTED',
  DISCONNECTED: 'DISCONNECTED',
  DEGRADED: 'DEGRADED',
} as const;

export interface TelemetryIngestRequest {
  node_id?: string;
  bridge_id?: string;
  /** @minimum 1 */
  sampling_rate: number;
  rms_g: number;
  peak_to_peak_g: number;
  fft_peak_hz: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  anomaly_score: number;
  status: TelemetryStatus;
  rssi_dbm: number;
  waveform_samples: number[];
}

export type TelemetryRecord = TelemetryIngestRequest & {
  id: number;
  node_id: string;
  bridge_id: string;
  timestamp: string;
};

export interface LiveVibration {
  node_id: string;
  sampling_rate: number;
  rms_g: number;
  peak_to_peak_g: number;
  fft_peak_hz: number;
  waveform: number[];
  anomaly_score: number;
  status: TelemetryStatus;
  timestamp: string;
}

export type DashboardSummaryMode = typeof DashboardSummaryMode[keyof typeof DashboardSummaryMode];


export const DashboardSummaryMode = {
  LIVE: 'LIVE',
  DEMO: 'DEMO',
} as const;

export interface DashboardSummary {
  bridge_health: number;
  tinyml_status: TelemetryStatus;
  anomaly_score: number;
  lora_network_status: LoraStatus;
  active_nodes: string;
  rssi_dbm: number;
  active_alerts_count: number;
  mode: DashboardSummaryMode;
  scenario: Scenario;
  last_updated: string;
}

export interface RecentEvent {
  id: number;
  event_type: string;
  node_id: string;
  bridge_id: string;
  severity: EventSeverity;
  anomaly_score?: number | null;
  message?: string | null;
  timestamp: string;
}

export interface NetworkPosture {
  node_id: string;
  lora_status: LoraStatus;
  rssi_dbm: number;
  active_nodes_count: number;
  total_nodes_count: number;
  packet_delivery_pct: number;
  gateway_latency_ms: number;
  bridge_health_pct: number;
  updated_at: string;
}

export interface ScenarioRequest {
  scenario: Scenario;
}

export type ModeRequestMode = typeof ModeRequestMode[keyof typeof ModeRequestMode];


export const ModeRequestMode = {
  LIVE: 'LIVE',
  DEMO: 'DEMO',
} as const;

export interface ModeRequest {
  mode: ModeRequestMode;
}

/**
 * Invalid request
 */
export type BadRequestResponse = ErrorResponse;

export type GetRecentEventsParams = {
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};