# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * GuardianBridge Lite telemetry and structural health monitoring API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryMode } from './dashboardSummaryMode';
import type { LoraStatus } from './loraStatus';
import type { Scenario } from './scenario';
import type { TelemetryStatus } from './telemetryStatus';

export interface DashboardSummary {
  bridge_health: number;
  tinyml_status: TelemetryStatus;
  anomaly_score: number;
  lora_network_status: LoraStatus;
  active_nodes: string;
  rssi_dbm: number;
  active_alerts_count: number;
  mode: DashboardSummaryMode;
  scenario: Scenario;
  last_updated: Date;
}