# lib/api-zod/src/generated/types/loraStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * GuardianBridge Lite telemetry and structural health monitoring API
 * OpenAPI spec version: 0.1.0
 */

export type LoraStatus = typeof LoraStatus[keyof typeof LoraStatus];


export const LoraStatus = {
  CONNECTED: 'CONNECTED',
  DISCONNECTED: 'DISCONNECTED',
  DEGRADED: 'DEGRADED',
} as const;