# lib/api-zod/src/generated/types/modeRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * GuardianBridge Lite telemetry and structural health monitoring API
 * OpenAPI spec version: 0.1.0
 */
import type { ModeRequestMode } from './modeRequestMode';

export interface ModeRequest {
  mode: ModeRequestMode;
}