# lib/api-zod/src/generated/types/networkPosture.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * GuardianBridge Lite telemetry and structural health monitoring API
 * OpenAPI spec version: 0.1.0
 */
import type { LoraStatus } from './loraStatus';

export interface NetworkPosture {
  node_id: string;
  lora_status: LoraStatus;
  rssi_dbm: number;
  active_nodes_count: number;
  total_nodes_count: number;
  packet_delivery_pct: number;
  gateway_latency_ms: number;
  bridge_health_pct: number;
  updated_at: Date;
}