# lib/api-zod/src/generated/types/recentEvent.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * GuardianBridge Lite telemetry and structural health monitoring API
 * OpenAPI spec version: 0.1.0
 */
import type { EventSeverity } from './eventSeverity';

export interface RecentEvent {
  id: number;
  event_type: string;
  node_id: string;
  bridge_id: string;
  severity: EventSeverity;
  anomaly_score?: number | null;
  message?: string | null;
  timestamp: Date;
}