# lib/api-zod/src/generated/types/scenario.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * GuardianBridge Lite telemetry and structural health monitoring API
 * OpenAPI spec version: 0.1.0
 */

export type Scenario = typeof Scenario[keyof typeof Scenario];


export const Scenario = {
  Normal_Operation: 'Normal Operation',
  Increased_Vibration: 'Increased Vibration',
  Loose_Joint: 'Loose Joint',
  Excess_Load: 'Excess Load',
} as const;