# lib/api-zod/src/generated/types/scenarioRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * GuardianBridge Lite telemetry and structural health monitoring API
 * OpenAPI spec version: 0.1.0
 */
import type { Scenario } from './scenario';

export interface ScenarioRequest {
  scenario: Scenario;
}