# lib/api-zod/src/generated/types/telemetryIngestRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * GuardianBridge Lite telemetry and structural health monitoring API
 * OpenAPI spec version: 0.1.0
 */
import type { TelemetryStatus } from './telemetryStatus';

export interface TelemetryIngestRequest {
  node_id?: string;
  bridge_id?: string;
  /** @minimum 1 */
  sampling_rate: number;
  rms_g: number;
  peak_to_peak_g: number;
  fft_peak_hz: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  anomaly_score: number;
  status: TelemetryStatus;
  rssi_dbm: number;
  waveform_samples: number[];
}