# lib/api-zod/src/generated/types/telemetryRecord.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * GuardianBridge Lite telemetry and structural health monitoring API
 * OpenAPI spec version: 0.1.0
 */
import type { TelemetryIngestRequest } from './telemetryIngestRequest';

export type TelemetryRecord = TelemetryIngestRequest & {
  id: number;
  node_id: string;
  bridge_id: string;
  timestamp: Date;
};