# lib/api-zod/src/generated/types/telemetryStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * GuardianBridge Lite telemetry and structural health monitoring API
 * OpenAPI spec version: 0.1.0
 */

export type TelemetryStatus = typeof TelemetryStatus[keyof typeof TelemetryStatus];


export const TelemetryStatus = {
  NORMAL: 'NORMAL',
  WARNING: 'WARNING',
  ANOMALY: 'ANOMALY',
  CRITICAL: 'CRITICAL',
} as const;